moved geometric sensitivity utility to core.
