+++ applications/AdjointFluidApplication/CMakeLists.txt
@@ -13,8 +13,6 @@ set( KRATOS_ADJOINT_FLUID_APPLICATION_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/custom_python/add_custom_processes_to_python.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/custom_python/adjoint_fluid_python_application.cpp 
     ${CMAKE_CURRENT_SOURCE_DIR}/custom_elements/vms_adjoint_element.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/custom_utilities/geometrical_sensitivity_utility.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/tests/test_geometrical_sensitivity_utility.cpp
     )
 
 find_package(HDF5 REQUIRED COMPONENTS CXX)
