+++ applications/AdjointFluidApplication/tests/run_cpp_unit_tests.py
@@ -1,4 +1,4 @@
 from KratosMultiphysics import *
-from KratosMultiphysics.AdjointFluidApplication import *
+#from KratosMultiphysics.AdjointFluidApplication import *
 Tester.SetVerbosity(Tester.Verbosity.TESTS_OUTPUTS)
 Tester.RunTestSuite("KratosSensitivityTestSuite")
