+++ kratos/utilities/geometrical_sensitivity_utility.h
